+++ tests/bootstrap.php
@@ -6,20 +6,35 @@ namespace
 
     // hostname for these servers will be in env vars
     // if it is not, default to localhost
-    if (!getenv('host-couchbase')) {
-        putenv('host-couchbase=127.0.0.1');
+    if (!getenv('couchbase-host')) {
+        putenv('couchbase-host=127.0.0.1');
     }
-    if (!getenv('host-memcached')) {
-        putenv('host-memcached=127.0.0.1');
+    if (!getenv('couchbase-port')) {
+        putenv('couchbase-port=11210');
     }
-    if (!getenv('host-mysql')) {
-        putenv('host-mysql=127.0.0.1');
+    if (!getenv('memcached-host')) {
+        putenv('memcached-host=127.0.0.1');
     }
-    if (!getenv('host-postgresql')) {
-        putenv('host-postgresql=127.0.0.1');
+    if (!getenv('memcached-port')) {
+        putenv('memcached-port=11211');
     }
-    if (!getenv('host-redis')) {
-        putenv('host-redis=127.0.0.1');
+    if (!getenv('mysql-host')) {
+        putenv('mysql-host=127.0.0.1');
+    }
+    if (!getenv('mysql-port')) {
+        putenv('mysql-port=3306');
+    }
+    if (!getenv('postgresql-host')) {
+        putenv('postgresql-host=127.0.0.1');
+    }
+    if (!getenv('postgresql-port')) {
+        putenv('postgresql-port=5432');
+    }
+    if (!getenv('redis-host')) {
+        putenv('redis-host=127.0.0.1');
+    }
+    if (!getenv('redis-port')) {
+        putenv('redis-port=6379');
     }
 
     // compatibility for when cache/integration-tests are run with PHPUnit>=6.0